#8-split code
