+++ pages/index.tsx
@@ -2,55 +2,19 @@ import { Box, Button, Card, Grid, TextField, Typography } from "@mui/material";
 import { useAppSelector } from "../redux/redux-hook";
 import { selectUser } from "../redux/slices/userSilce";
 import Avatar from "@mui/material/Avatar";
-import {
-  IGetSocialAccountsResponse,
-  IGraphQLError,
-  ISocialAcount,
-} from "../types/graphql.respose";
+import { ISocialAcount } from "../types/graphql.respose";
 import { useEffect, useState } from "react";
-import client from "../apollo-client";
-import { GET_ACCOUNTS_QUERY } from "../graphql/home-page";
-import toast from "react-hot-toast";
 import { SocialType } from "../constance/enum";
+import { getAccounts } from "../utils/home";
 
 export default function Home() {
   const user = useAppSelector(selectUser);
   const [socialAccounts, setSocialAccounts] = useState<ISocialAcount[]>([]);
 
   const fetchListAccounts = async () => {
-    const token = localStorage.getItem("accessToken");
-
-    try {
-      const { data } = await client.query<IGetSocialAccountsResponse>({
-        query: GET_ACCOUNTS_QUERY,
-        context: {
-          headers: {
-            authorization: token ? `Bearer ${token}` : "",
-          },
-        },
-      });
-
-      setSocialAccounts(data.getAccountsLinked);
-      // eslint-disable-next-line @typescript-eslint/no-explicit-any
-    } catch (error: any) {
-      const { graphQLErrors } = error;
-
-      if (graphQLErrors) {
-        const { statusCode } = graphQLErrors[0] as IGraphQLError;
-        console.log("statusCode: ", statusCode);
-
-        switch (statusCode) {
-          case 400:
-            toast.error("You are unauthrorized");
-            break;
-
-          default:
-            toast.error("Something went wrong");
-            break;
-        }
-      } else {
-        toast.error("Something went wrong");
-      }
+    const respose = await getAccounts();
+    if (respose) {
+      setSocialAccounts(respose);
     }
   };
